==> python/0_revised_initModel.ipynb <==
import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
from sklearn.preprocessing import LabelEncoder, MinMaxScaler
from sklearn.random_projection import GaussianRandomProjection
from sklearn.random_projection import SparseRandomProjection
##Add decomposed components: PCA / ICA etc.
from sklearn.decomposition import PCA, FastICA
from sklearn.decomposition import TruncatedSVD
import xgboost as xgb
from sklearn.metrics import r2_score
from sklearn.model_selection import StratifiedKFold

## 1. Load

# load data
dt_train_raw = pd.read_csv("../../data/Mercedes_Benz_Greener_Manufacturing/raw/train.csv")
dt_test_raw = pd.read_csv("../../data/Mercedes_Benz_Greener_Manufacturing/raw/test.csv")

dt_train_raw.head()

print(dt_train_raw.shape, dt_test_raw.shape)

## 2. Transform

# # marke train and test set
# dt_train_raw.loc[:, "IsTrainTest"] = pd.Series("train", index = dt_train_raw.index)
# dt_test_raw.loc[:, "IsTrainTest"] = pd.Series("test", index = dt_test_raw.index)

# # change test index and add y
# dt_test_raw.index = dt_test_raw.index + max(dt_train_raw.index) + 1
# dt_test_raw.loc[:,"y"] = pd.Series([0.0] * dt_test_raw.shape[0], index = dt_test_raw.index)

# # concat
# dt_test_raw = dt_test_raw[dt_train_raw.columns.values]
# dt_all_raw = pd.concat([dt_train_raw, dt_test_raw])

dt_all_raw.shape

## 3. Encode

# cols_cat
cols_cat = dt_all_raw.drop("IsTrainTest", axis = 1).select_dtypes(include = ['object']).columns.values
dt_cat = dt_all_raw[cols_cat]

for c in cols_cat:
    x = list(set(dt_cat[c].values))
    x.sort()
    x.sort(key = len)
    dt_labelEncode_c = pd.DataFrame({"Encode_" + c: [i for i in range(1, (len(x) + 1))]
                                     , c: x})
    
    dt_cat = pd.merge(dt_cat, dt_labelEncode_c, on = c)
    

dt_cat.head()

dt_all_raw = dt_all_raw.drop(cols_cat, axis = 1).join(dt_cat.filter(regex = "Encode"))

dt_all_raw.head()

## 3. Decomposition

# # range
# pp_range = MinMaxScaler()
# mx_range = pp_range.fit_transform(dt_all_raw.drop(["y", "IsTrainTest"], axis = 1))
# dt_all_raw = dt_all_raw[["y", "IsTrainTest"]].join(pd.DataFrame(mx_range, columns = dt_all_raw.drop(["y", "IsTrainTest"], axis = 1).columns.values))
# dt_all_raw.shape

dt_all_raw.head()

n_comp = 12

# tSVD
tsvd = TruncatedSVD(n_components=n_comp, random_state=420)
tsvd_results_all = tsvd.fit_transform(dt_all_raw.drop(["y", "IsTrainTest"], axis=1))

# PCA
pca = PCA(n_components=n_comp, random_state=420)
pca2_results_all = pca.fit_transform(dt_all_raw.drop(["y", "IsTrainTest"], axis=1))

# ICA
ica = FastICA(n_components=n_comp, random_state=420)
ica2_results_all = ica.fit_transform(dt_all_raw.drop(["y", "IsTrainTest"], axis=1))

# GRP
grp = GaussianRandomProjection(n_components=n_comp, eps=0.1, random_state=420)
grp_results_all = grp.fit_transform(dt_all_raw.drop(["y", "IsTrainTest"], axis=1))

# SRP
srp = SparseRandomProjection(n_components=n_comp, dense_output=True, random_state=420)
srp_results_all = srp.fit_transform(dt_all_raw.drop(["y", "IsTrainTest"], axis=1))

# Append decomposition components to datasets
for i in range(1, n_comp+1):
    dt_all_raw['pca_' + str(i)] = tsvd_results_all[:,i-1]

    dt_all_raw['ica_' + str(i)] = pca2_results_all[:,i-1]

    dt_all_raw['tsvd_' + str(i)] = ica2_results_all[:,i-1]

    dt_all_raw['grp_' + str(i)] = grp_results_all[:,i-1]

    dt_all_raw['srp_' + str(i)] = srp_results_all[:,i-1]

y_train = dt_all_raw.loc[dt_all_raw["IsTrainTest"] == "train"]["y"]
y_mean = np.mean(y_train)

dt_all_raw.head()

## 4. Feature Engineering

### 4.1 Outlier Maker

dt_featEng = pd.read_csv("../../data/Mercedes_Benz_Greener_Manufacturing/data/dt_all_eng.csv")

dt_featEng.filter(regex = "Outlier")

## 5. xgboost

# r_2 for xgboost
def r_2(preds, dtrain):
    labels = dtrain.get_label()
    return 'score', r2_score(labels, preds)

params_xgb = {
    'n_trees': 520,
    'eta': 0.0045,
    'max_depth': 4,
    'subsample': 0.98,
    'objective': 'reg:linear',
    'eval_metric': 'rmse',
    'base_score': y_mean, # base prediction = mean(target)
    'silent': 1
}

# xgbDmatrix
dtrain = xgb.DMatrix(dt_all_raw.loc[dt_all_raw["IsTrainTest"] == "train"].drop(["y", "IsTrainTest"], axis=1), y_train)
dtest = xgb.DMatrix(dt_all_raw.loc[dt_all_raw["IsTrainTest"] == "test"].drop(["y", "IsTrainTest"], axis=1))

# cv
cv_xgb = xgb.cv(params_xgb, dtrain
                , num_boost_round = 5000
                , nfold = 10
                , feval = r_2, maximize = True, early_stopping_rounds = 50
                , show_stdv = True, verbose_eval = 50, seed = 888)
