==> python/0_run.ipynb <==
# import
import warnings
warnings.filterwarnings('ignore')
import sys
sys.path.insert(0, "/media/noahhhhhh/dataScience/proj/competition/Mercedes_Benz_Greener_Manufacturing/python/utils/")
import pickle

import numpy as np
import pandas as pd
import gc

# utils
from clean import *
from encode import *
from featureEngineer import *
from model import *
from dups import *


########################### load ###########################
dt_all = pd.read_csv("/media/noahhhhhh/dataScience/proj/competition/data/Mercedes_Benz_Greener_Manufacturing/data/dt_all.csv")
dt_train_raw = pd.read_csv("/media/noahhhhhh/dataScience/proj/competition/data/Mercedes_Benz_Greener_Manufacturing/raw/train.csv")
dt_test_raw = pd.read_csv("/media/noahhhhhh/dataScience/proj/competition/data/Mercedes_Benz_Greener_Manufacturing/raw/test.csv")

# ids
ids_train = dt_train_raw.ID.values
ids_test = dt_test_raw.ID.values

# cols
cols_cat = dt_all.drop("ID", axis = 1).select_dtypes(include = ["object"]).columns.values.tolist()
cols_bin = dt_all.filter(regex = "^X").columns.values.tolist()
cols_bin = [x for x in cols_bin if x not in cols_cat]

# ## X. Model

dt_all_train = dt_all[dt_all.ID.isin(ids_train)]
dt_all_test = dt_all[dt_all.ID.isin(ids_test)]

print(dt_all_train.shape, dt_all_test.shape)


# # ### X.1. xgb
# with outlier
#params_cols_dict_best_xgb_withOutlier = featureSelectRun("xgb", dt_all, ids_train, ids_test, cols_cat, cols_bin, rounds = 200)
#pickle.dump(params_cols_dict_best_xgb_withOutlier, open( "/media/noahhhhhh/dataScience/proj/competition/data/Mercedes_Benz_Greener_Manufacturing/data/params_cols_dict_best_xgb_withOutlier.pkl", "wb"))

# with outlier and avg duplicates
dt_withOutlier_aggDupsMedian = dups("min", dt_all, cols_cat, cols_bin)

dt_all_train = dt_all[dt_all["y"].notnull()]
dt_all_test = dt_all[~dt_all["y"].notnull()]

# folds
k = 10
bin_y = pd.qcut(dt_all_train.y.values, k, labels = [i for i in range(1, k + 1)]).astype("int64")
dt_all_train["Fold"] = bin_y

from sklearn.model_selection import StratifiedKFold
skf = StratifiedKFold(n_splits = k, shuffle = True, random_state = 888)
dt_id_folds = pd.DataFrame()
for i, (ind_train, ind_valid) in enumerate(skf.split(dt_all_train, bin_y)):
    dt_id_folds = pd.concat([dt_id_folds
                            , pd.DataFrame({"ID": dt_all_train.iloc[ind_valid].ID.values
                                           , "Fold": i + 1})])

pd.merge(dt_all_train.drop("Fold", axis = 1), dt_id_folds, on = "ID")

dt_all_train.columns.values

dt_withOutlier_aggDupsMedian.shape

dt_withOutlier_aggDupsMedian.head()

featureSelectRun("xgb", dt_withOutlier_aggDupsMedian, ids_train, ids_test, cols_cat, cols_bin, rounds = 100, autoFolds = True)
