==> python/1_transform.ipynb <==
import pandas as pd
import pickle

## 1. Load

# load data
dt_train_raw = pd.read_csv("../../data/Mercedes_Benz_Greener_Manufacturing/raw/train.csv")
dt_test_raw = pd.read_csv("../../data/Mercedes_Benz_Greener_Manufacturing/raw/test.csv")
# cols
cols_cat = dt_train_raw.select_dtypes(include = ['object']).columns.values
cols_cat = cols_cat.tolist()

print(dt_train_raw.shape, dt_test_raw.shape)

## 2. Transform

# change test index and add y
dt_test_raw["y"] = 0.0

# concat
dt_test_raw = dt_test_raw[dt_train_raw.columns.values]
dt_all_raw = pd.concat([dt_train_raw, dt_test_raw])

dt_all_raw.shape

## 3. Save

IDs_train = dt_train_raw.ID.values
IDs_test = dt_test_raw.ID.values

# data
dt_all_raw.to_csv("../../data/Mercedes_Benz_Greener_Manufacturing/data/dt_all_transformed.csv", index = False)
# id
pickle.dump(IDs_train, open( "../../data/Mercedes_Benz_Greener_Manufacturing/data/IDs_train.pkl", "wb"))
pickle.dump(IDs_test, open( "../../data/Mercedes_Benz_Greener_Manufacturing/data/IDs_test.pkl", "wb"))
# cols
pickle.dump(cols_cat, open( "../../data/Mercedes_Benz_Greener_Manufacturing/data/cols_cat.pkl", "wb"))
