==> python/4_featEng.ipynb <==
import pandas as pd
import pickle
import numpy as np
from scipy import stats
from sklearn import preprocessing, decomposition, manifold
from sklearn import random_projection
import operator

## 4. Feature Engineering

# dt_all_eng
dt_all_eng = pd.read_csv("../../data/Mercedes_Benz_Greener_Manufacturing/data/dt_all_encoded.csv")
# cols
cols_raw = pickle.load(open("../../data/Mercedes_Benz_Greener_Manufacturing/data/cols_raw.pkl", "rb"))
cols_cat = pickle.load(open("../../data/Mercedes_Benz_Greener_Manufacturing/data/cols_cat.pkl", "rb"))
# IDs
IDs_train = pickle.load(open("../../data/Mercedes_Benz_Greener_Manufacturing/data/IDs_train.pkl", "rb"))
IDs_test = pickle.load(open("../../data/Mercedes_Benz_Greener_Manufacturing/data/IDs_test.pkl", "rb"))

dt_all_eng.shape

### 4.1 Outlier marker

def getOutlierMarker(dt, cols_type, id_outlier = 1770):
    
    if cols_type == "Cat":
        cols = ['X0', 'X1', 'X2', 'X3', 'X4', 'X5', 'X6', 'X8']
    elif cols_type == "Bin":
        cols = cols_raw
    elif cols_type == "All":
        cols = ['X0', 'X1', 'X2', 'X3', 'X4', 'X5', 'X6', 'X8'] + list(cols_raw)
    else:
        cols = [cols_type]
    
    # outlier
    dt_outlier = dt[cols].loc[dt["ID"] == id_outlier]
    # calc int_outlierMarker
    dt_outlierMaker = dt_all_eng[["ID"]]
    for col in cols:
        dt_outlierMaker_id_col = pd.DataFrame()
        for id in dt.ID.values:
            is_equal = dt.loc[dt["ID"] == id][col].values == dt_outlier[col].values
            dt_outlierMaker_id_col = pd.concat([dt_outlierMaker_id_col
                                        , pd.DataFrame({"ID": id
                                           , col + "_is_equal": is_equal.astype("int64")})])
        
        dt_outlierMaker = pd.merge(dt_outlierMaker, dt_outlierMaker_id_col, on = "ID")
        
    sum_outlierMaker = dt_outlierMaker.filter(regex = "_is_equal").sum(axis = 1)
    dt["FeatEng_OutlierMaker_" + str(id_outlier) + "_" + cols_type] = sum_outlierMaker
    
    return dt

#### 4.1.1 OutlierMarker_Cat

# dt_all_eng = getOutlierMarker(dt_all_eng, cols_type = "Cat")

#### 4.1.2 OutlierMarker_Int

# dt_all_eng = getOutlierMarker(dt_all_eng, cols_type = "Bin")

#### 4.1.3 OutlierMarker_All

# dt_all_eng = getOutlierMarker(dt_all_eng, cols_type = "All")

#### 4.1.4 OutlierMarker_X0

# dt_all_eng = getOutlierMarker(dt_all_eng, cols_type = "X0")

dt_all_eng.head()

### 4.2 Sum of binary cols

#### 4.2.1 Sum of all binary cols

# sum_binary_all = dt_all_eng[cols_raw].sum(axis = 1)
# dt_all_eng["FeatEng_SumBin_All"] = sum_binary_all

#### 4.2.2 Sum of correlation-important binary cols

def corBin_Contin(dt, cols, method = "pointbiserialr"):
    dt_binary_pointbiserialr = pd.DataFrame()
    for col in cols:
        if method == "spearmanr":
            cor_pb = stats.spearmanr(dt[col].values, dt.y.values)
        else:
            cor_pb = stats.pointbiserialr(dt[col].values, dt.y.values)
        dt_binary_pointbiserialr = pd.concat([dt_binary_pointbiserialr
                                             , pd.DataFrame({"Col": col
                                                            , "Cor": np.abs([cor_pb.correlation])
                                                            , "P": [cor_pb.pvalue]})])
    return dt_binary_pointbiserialr.sort_values("Cor", ascending = False)

# dt_corBin_Cotin = corBin_Contin(dt_all_eng.loc[dt_all_eng["ID"].isin(IDs_train)], cols_raw)

# dt_corBin_Cotin_spear = corBin_Contin(dt_all_eng.loc[dt_all_eng["ID"].isin(IDs_train)], cols_raw, method = "spearmanr")

# cols_binary_cor_02_10 = dt_corBin_Cotin_spear.loc[dt_corBin_Cotin_spear["Cor"] >= .2]["Col"].values
# cols_binary_cor_01_02 = dt_corBin_Cotin_spear.loc[(dt_corBin_Cotin_spear["Cor"] >= .1) & (dt_corBin_Cotin_spear["Cor"] < .2)]["Col"].values
# cols_binary_cor_005_01 = dt_corBin_Cotin_spear.loc[(dt_corBin_Cotin_spear["Cor"] >= .05) & (dt_corBin_Cotin_spear["Cor"] < .1)]["Col"].values

# cols_binary_important_3 = dt_corBin_Cotin["Col"].values[:3]
# cols_binary_important_5 = dt_corBin_Cotin["Col"].values[:5]
# cols_binary_important_10 = dt_corBin_Cotin["Col"].values[:10]
# cols_binary_important_20 = dt_corBin_Cotin["Col"].values[:20]
# cols_binary_important_50 = dt_corBin_Cotin["Col"].values[:50]

# sum_binary_important_3 = dt_all_eng[cols_binary_important_3].sum(axis = 1)
# sum_binary_important_5 = dt_all_eng[cols_binary_important_5].sum(axis = 1)
# sum_binary_important_10 = dt_all_eng[cols_binary_important_10].sum(axis = 1)
# sum_binary_important_20 = dt_all_eng[cols_binary_important_20].sum(axis = 1)
# sum_binary_important_50 = dt_all_eng[cols_binary_important_50].sum(axis = 1)
# sum_binary_important_cor_02_10 = dt_all_eng[cols_binary_cor_02_10].sum(axis = 1)
# sum_binary_important_cor_01_02 = dt_all_eng[cols_binary_cor_01_02].sum(axis = 1)
# sum_binary_important_cor_005_01 = dt_all_eng[cols_binary_cor_005_01].sum(axis = 1)

# # SumBin_All
# dt_all_eng.loc[:, "FeatEng_SumBin_All"] = sum_binary_all
# # SumBin_Imp_X
# dt_all_eng.loc[:, "FeatEng_SumBin_Imp_3"] = sum_binary_important_3
# dt_all_eng.loc[:, "FeatEng_SumBin_Imp_5"] = sum_binary_important_5
# dt_all_eng.loc[:, "FeatEng_SumBin_Imp_10"] = sum_binary_important_10
# dt_all_eng.loc[:, "FeatEng_SumBin_Imp_20"] = sum_binary_important_20
# dt_all_eng.loc[:, "FeatEng_SumBin_Imp_50"] = sum_binary_important_50
# dt_all_eng.loc[:, "FeatEng_SumBin_Cor_02_10"] = sum_binary_important_cor_02_10
# dt_all_eng.loc[:, "FeatEng_SumBin_Cor_01_02"] = sum_binary_important_cor_01_02
# dt_all_eng.loc[:, "FeatEng_SumBin_Cor_005_01"] = sum_binary_important_cor_005_01

dt_all_eng.shape

### 4.3 Dimension reduction

# range
pp_range = preprocessing.StandardScaler()
mx_range = pp_range.fit_transform(dt_all_eng.drop(["y"] + cols_cat, axis = 1))
dt_all_eng_norm = pd.DataFrame(mx_range, columns = dt_all_eng.drop(["y"] + cols_cat, axis = 1).columns.values)
dt_all_eng_norm = dt_all_eng_norm.rename(columns = {"ID": "ID_Norm"})
dt_all_eng = dt_all_eng_norm.join(dt_all_eng.y).join(dt_all_eng.ID)

def featEng_dimRed(method, dt, n_component, name_feature):
        
    if method == "PCA":
        dr = decomposition.PCA(n_components = n_component * 2, random_state = 420)
    elif method == "ICA":
        dr = decomposition.FastICA(n_components = n_component, max_iter = 5000, tol = 0.001, random_state = 420)
    elif method == "SVD":
        dr = decomposition.TruncatedSVD(n_components = n_component, random_state = 420)
    elif method == "FA":
        dr = decomposition.FactorAnalysis(n_components = n_component, random_state = 420)
    elif method == "TSNE":
        dr = manifold.TSNE(n_components = n_component, random_state = 420)
    elif method == "GRP":
        dr = random_projection.GaussianRandomProjection(n_components = n_component, eps = 0.1, random_state = 420)
    elif method == "SRP":
        dr = random_projection.SparseRandomProjection(n_components = n_component, dense_output = True, random_state = 420)
    
    if name_feature == "Raw_Bin":
        cols = cols_raw
    elif name_feature == "Encoded_Cat":
        cols = dt.filter(regex = "Encode_").columns.values
    elif name_feature == "FeatEng":
        cols = dt.filter(regex = "FeatEng_").columns.values
    elif name_feature == "All":
        cols = dt.drop(["y"], axis = 1).filter(regex = "^((?!DR).)*$").columns.values
    else:
        cols = name_feature
        
    mx = dr.fit_transform(dt[cols])
    if method == "PCA":
        dt = dt.join(pd.DataFrame(mx, columns = ["DR_" + method + "_" + name_feature + "_" + str(i) for i in range(1, n_component * 2 + 1)]))
    else:
        dt = dt.join(pd.DataFrame(mx, columns = ["DR_" + method + "_" + name_feature + "_" + str(i) for i in range(1, n_component + 1)]))
        
    return dt

methods_dr = ["PCA", "ICA", "SVD", "FA", "GRP", "SRP"] # TSNE
for method in methods_dr:
#     dt_all_eng = featEng_dimRed(method, dt_all_eng, 20, "Raw_Bin")
#     dt_all_eng = featEng_dimRed(method, dt_all_eng, 20, "Encoded_Cat")
#     dt_all_eng = featEng_dimRed(method, dt_all_eng, 3, "FeatEng")
    dt_all_eng = featEng_dimRed(method, dt_all_eng, 50, "All")

dt_all_eng.shape

### 4.5 Feature Interaction

# dt_interest = dt_all_eng[cols_raw]
# dt_fi_2way = pd.DataFrame()
# seen = []
# for c1 in cols_raw:
#     for c2 in cols_raw:
#         if c1 != c2:
#             seen.append((c1, c2))
#             if (c2, c1) not in seen:
#                 isTrain = dt_all_eng["ID"].isin(IDs_train)
#                 val_c1 = dt_interest[c1].values.astype("int64")
#                 val_c2 = dt_interest[c2].values.astype("int64")
#                 y_train = dt_all_eng.y.values[isTrain]
#                 # XOR
#                 XOR = val_c1 ^ val_c2
#                 cor_c1 = abs(stats.spearmanr(val_c1[isTrain], y_train).correlation)
#                 cor_c2 = abs(stats.spearmanr(val_c2[isTrain], y_train).correlation)
#                 cor_XOR = abs(stats.spearmanr(XOR[isTrain], y_train).correlation)
#                 # constant
#                 if cor_XOR > .3:
#                     const = 1.1
#                 elif (cor_XOR > .2) & (cor_XOR <= .3):
#                     const = 1.3
#                 else:
#                     const = 1.5
#                 if cor_XOR > max(cor_c1, cor_c2) * const:
#                     dt_fi_2way.loc[:, "FeatEng_FI_2way_XOR_" + c1 + "_" + c2] = XOR

# dt_all_eng = dt_all_eng.join(dt_fi_2way)

dt_all_eng.shape

### 4.6 Save dt_all_eng

dt_all_eng.to_csv("../../data/Mercedes_Benz_Greener_Manufacturing/data/dt_all_eng.csv", index = False)
