==> python/5_preprocess.ipynb <==
import pandas as pd
from sklearn import preprocessing

## 5. Preprocess

dt_all_preprocess = pd.read_csv("../../data/Mercedes_Benz_Greener_Manufacturing/data/dt_all_eng.csv")

dt_id_y_istraintest = dt_all_preprocess[["ID", "y", "IsTrainTest"]]

### 5.1 Range

# range
pp_range = preprocessing.MinMaxScaler()
mx_range = pp_range.fit_transform(dt_all_preprocess.drop(["ID", "y", "IsTrainTest"], axis = 1))
dt_all_preprocess = pd.DataFrame(mx_range, columns = dt_all_preprocess.drop(["ID", "y", "IsTrainTest"], axis = 1).columns.values)
dt_all_preprocess.shape

dt_all_preprocess = dt_all_preprocess.join(dt_id_y_istraintest)

# csv
dt_all_preprocess.to_csv("../../data/Mercedes_Benz_Greener_Manufacturing/data/dt_all_preprocess.csv", index = False)
