==> python/6_model_nn.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.model_selection import cross_val_score, KFold, train_test_split, GridSearchCV, StratifiedKFold
from sklearn.feature_selection import SelectFromModel
from sklearn.metrics import r2_score, mean_squared_error

# Keras is a deep learning library that wraps the efficient numerical libraries Theano and TensorFlow.
# It provides a clean and simple API that allows you to define and evaluate deep learning models in just a few lines of code.from keras.models import Sequential, load_model
from keras.models import Sequential, load_model
from keras.layers import Dense, Dropout, BatchNormalization, Activation
from keras.wrappers.scikit_learn import KerasRegressor
from keras.callbacks import EarlyStopping, ModelCheckpoint
# define custom R2 metrics for Keras backend
from keras import backend as K
# to tune the NN
from keras.constraints import maxnorm
from keras.optimizers import SGD, Adam
from keras.layers.advanced_activations import PReLU
from keras.regularizers import l2
import h5py

# define path to save model
import os
model_path = "../../data/Mercedes_Benz_Greener_Manufacturing/model/model_nn.h5"

## 6. Model

# data
dt_all = pd.read_csv("/media/noahhhhhh/dataScience/proj/competition/data/Mercedes_Benz_Greener_Manufacturing/data/dt_all.csv")
dt_train_raw = pd.read_csv("/media/noahhhhhh/dataScience/proj/competition/data/Mercedes_Benz_Greener_Manufacturing/raw/train.csv")
dt_test_raw = pd.read_csv("/media/noahhhhhh/dataScience/proj/competition/data/Mercedes_Benz_Greener_Manufacturing/raw/test.csv")

cols_cat = dt_all.drop("ID", axis = 1).select_dtypes(include = ["object"]).columns.values.tolist()
cols_bin = dt_all.filter(regex = "^X").columns.values.tolist()
cols_bin = [x for x in cols_bin if x not in cols_cat]


dt_test = dt_all[~dt_all["y"].notnull()]
dt_train = dt_all[dt_all["y"].notnull()]
dt_train_dups = dt_train[dt_train[cols_cat + cols_bin].duplicated(keep = False) == True]
dt_train_nonDups = dt_train[dt_train[cols_cat + cols_bin].duplicated(keep = False) == False]

dt_dups_agg = pd.DataFrame({"y" : dt_train_dups.groupby(cols_cat + cols_bin)["y"].apply(lambda x: sorted(x)[math.ceil(len(x) / 2)])}).reset_index()

dt_dups_agg = pd.DataFrame({"y" : dt_train_dups.groupby(cols_cat + cols_bin)["y"].apply(lambda x: sorted(x)[math.ceil(len(x) / 2)])}).reset_index()
dt_nonDups = dt_train_dups[dt_train_dups["y"].isin(dt_dups_agg["y"])]
dt_nonDup_agg = pd.DataFrame({"ID" : dt_nonDups.groupby(cols_cat + cols_bin)["ID"].max()}).reset_index()
dt_nonDups_train = dt_train_dups[dt_train_dups["ID"].isin(dt_nonDup_agg["ID"])]

import math
x = np.array([1, 3, 2, 4])
sorted(x)[math.ceil(len(x) / 2)]

def median_noah(series):
    if len(series) % 2 == 0:
        return(sorted(series)[math.ceil(len(series) / 2)])
    else:
        return(np.median(series))
        

median_noah([10, 9, 12, 11])

def dups(method, dt_all, cols_cat, cols_bin):

   
    dt_test = dt_all[~dt_all["y"].notnull()]
    dt_train = dt_all[dt_all["y"].notnull()]
    dt_train_dups = dt_train[dt_train[cols_cat + cols_bin].duplicated(keep = False) == True]
    dt_train_nonDups = dt_train[dt_train[cols_cat + cols_bin].duplicated(keep = False) == False]

    if method == "max":
        dt_dups_agg = pd.DataFrame({"y" : dt_train_dups.groupby(cols_cat + cols_bin)["y"].max()}).reset_index()
        dt_nonDups = dt_train_dups[dt_train_dups["y"].isin(dt_dups_agg["y"])]
        dt_nonDup_agg = pd.DataFrame({"ID" : dt_nonDups.groupby(cols_cat + cols_bin)["ID"].max()}).reset_index()
        dt_nonDups_train = dt_train_dups[dt_train_dups["ID"].isin(dt_nonDup_agg["ID"])]
    elif method == "min":
        dt_dups_agg = pd.DataFrame({"y" : dt_train_dups.groupby(cols_cat + cols_bin)["y"].min()}).reset_index()
        dt_nonDups = dt_train_dups[dt_train_dups["y"].isin(dt_dups_agg["y"])]
        dt_nonDup_agg = pd.DataFrame({"ID" : dt_nonDups.groupby(cols_cat + cols_bin)["ID"].max()}).reset_index()
        dt_nonDups_train = dt_train_dups[dt_train_dups["ID"].isin(dt_nonDup_agg["ID"])]
    elif method == "median":
        dt_dups_agg = pd.DataFrame({"y" : dt_train_dups.groupby(cols_cat + cols_bin)["y"].apply(lambda x: sorted(x)[math.ceil(len(x) / 2)])}).reset_index()
        dt_nonDups = dt_train_dups[dt_train_dups["y"].isin(dt_dups_agg["y"])]
        dt_nonDup_agg = pd.DataFrame({"ID" : dt_nonDups.groupby(cols_cat + cols_bin)["ID"].max()}).reset_index()
        dt_nonDups_train = dt_train_dups[dt_train_dups["ID"].isin(dt_nonDup_agg["ID"])]
    
    return(pd.concat([dt_nonDups_train, dt_train_nonDups, dt_test]))

dups("median", dt_all, cols_cat, cols_bin)

# ids
ids_train = dt_train_raw.ID.values
ids_test = dt_test_raw.ID.values

# r_2 for nn
def r2_keras(y_true, y_pred):
    SS_res =  K.sum(K.square( y_true - y_pred )) 
    SS_tot = K.sum(K.square( y_true - K.mean(y_true) ) ) 
    return ( 1 - SS_res/(SS_tot + K.epsilon()) )

# X, y, ID
X_train_all = dt_all[dt_all["ID"].isin(ids_train)].as_matrix()
X_test = dt_all[dt_all["ID"].isin(ids_test)].as_matrix()
y_train_all = dt_all[dt_all["ID"].isin(ids_train)].y.values

# keras
def model(dropout_level = 0.25, activation = 'tanh'):
    model = Sequential()
    model.add(Dense(1024, input_dim=input_dims, kernel_initializer="he_normal"))
    model.add(BatchNormalization())
    model.add(Activation(activation))
    model.add(Dropout(dropout_level))
    
    model.add(Dense(1024, kernel_initializer="he_normal"))
    model.add(BatchNormalization())
    model.add(Activation(activation))
    model.add(Dropout(dropout_level))
    
    model.add(Dense(1024, kernel_initializer="he_normal"))
    model.add(BatchNormalization())
    model.add(Activation(activation))
    model.add(Dropout(dropout_level))
    
    model.add(Dense(1024, kernel_initializer="he_normal"))
    model.add(BatchNormalization())
    model.add(Activation(activation))
    model.add(Dropout(dropout_level))
    
    model.add(Dense(1024, kernel_initializer="he_normal"))
    model.add(BatchNormalization())
    model.add(Activation(activation))
    model.add(Dropout(dropout_level))
    
    model.add(Dense(768, kernel_initializer="he_normal"))
    model.add(BatchNormalization())
    model.add(Activation(activation))
    model.add(Dropout(dropout_level))
    
    model.add(Dense(768, kernel_initializer="he_normal"))
    model.add(BatchNormalization())
    model.add(Activation(activation))
    model.add(Dropout(dropout_level))
    
    model.add(Dense(768, kernel_initializer="he_normal"))
    model.add(BatchNormalization())
    model.add(Activation(activation))
    model.add(Dropout(dropout_level))

    model.add(Dense(768, kernel_initializer="he_normal", kernel_regularizer = l2(1.e-5)))
    model.add(BatchNormalization())
    model.add(Activation(activation))
    model.add(Dropout(dropout_level))
    
    model.add(Dense(1, activation='linear'))
    
    model.compile(loss = "mean_squared_error", # one may use 'mean_absolute_error' as alternative
                  optimizer = "adam",
                  metrics = [r2_keras, "accuracy"] # you can add several if needed
                 )
    
    # Visualize NN architecture
    print(model.summary())
    return model

input_dims = X_train_all.shape[1]
estimator = KerasRegressor(
    build_fn = model, 
    nb_epoch = 300, 
    batch_size = 35,
    verbose = 1
)

# prepare callbacks
callbacks = [
    EarlyStopping(
        monitor = 'val_loss', 
        patience = 20,
        verbose = 1),
    ModelCheckpoint(
        model_path, 
        monitor = 'val_loss', 
        save_best_only = True, 
        verbose = 0)
]

#K-FOLD
n_splits = 5
kf = KFold(n_splits = n_splits)
kf.get_n_splits(X_train_all)

predictions = np.zeros((X_train_all.shape[0], n_splits))
score = 0

for fold, (ind_train, ind_valid) in enumerate(kf.split(X_train_all)):

    X_train, X_valid = X_train_all[ind_train, :], X_train_all[ind_valid, :]
    y_train, y_valid = y_train_all[ind_train], y_train_all[ind_valid]

    # fit estimator
    history = estimator.fit(
        X_train, 
        y_train, 
        batch_size = 64,
        epochs = 500,
        validation_data = (X_valid, y_valid),
        verbose = 2,
        callbacks = callbacks,
        shuffle = True
    )
    
    if os.path.isfile(model_path):
        history.model.load_weights(model_path)
    
    pred = history.model.predict(X_valid)
    
    score_fold = r2_score(y_valid, pred)
    score += score_fold

    print('Fold %d: Score %f'%(fold, score_fold))


score /= n_splits

print('=====================')

print( 'Final Score %f'%score)

print('=====================')
