==> python/7_experiment.ipynb <==
import pandas as pd
import xgboost as xgb
import numpy as np
from scipy.stats import uniform as sp_rand
from scipy.stats import randint as sp_randint
import random
from sklearn.metrics import r2_score, make_scorer
from sklearn.model_selection import StratifiedKFold, GroupKFold, KFold, train_test_split
from sklearn.model_selection import GridSearchCV, RandomizedSearchCV
import matplotlib.pyplot as plt
import pickle

## 7. Experiment

# data
dt_model = pd.read_csv("../../data/Mercedes_Benz_Greener_Manufacturing/data/dt_all_preprocess.csv")
params_xgb = pickle.load(open("../../data/Mercedes_Benz_Greener_Manufacturing/data/params_xgb_05.pkl", "rb"))
cols_raw = pickle.load(open("../../data/Mercedes_Benz_Greener_Manufacturing/data/cols_raw.pkl", "rb"))

dt_model.shape

params_xgb = {'booster': 'dart',
 'colsample_bylevel': 0.8,
 'colsample_bytree': 0.8,
 'gamma': 0.1,
 'learning_rate': 0.002,
 'max_depth': 2,
 'min_child_weight': 1,
 'n_estimators': 3000,
 'reg_alpha': 0.01,
 'reg_lambda': 0.01,
 'subsample': 0.88}

# r_2 for xgboost
def r_2(preds, dtrain):
    labels = dtrain.get_label()
    return 'score', r2_score(labels, preds)

# make scorer_r2
scorer_r2 = make_scorer(r2_score)

### Feature set selection

# feature set select
list_cols = ["single_train", "single_test", "dup_train", "dup_test", "Is"
             , "Encode_ohe", "Encode_TargetMean", "Encode_Freq", "Encode_Binary", "Encode_Ordinal"
             , "FeatEng_OutlierMarker", "FeatEng_SumBin"
             , "DR_PCA_Raw_Bin", "DR_PCA_Encoded_Cat", "DR_PCA_FeatEng", "DR_PCA_All"
             , "DR_ICA_Raw_Bin", "DR_ICA_Encoded", "DR_ICA_FeatEng", "DR_ICA_All"
             , "DR_SVD_Raw_Bin", "DR_SVD_Encoded", "DR_SVD_FeatEng", "DR_SVD_All"
             , "DR_FA_Raw_Bin", "DR_FA_Encoded", "DR_FA_FeatEng", "DR_FA_All"
             , "DR_TSNE_Raw_Bin", "DR_TSNE_Encoded", "DR_TSNE_FeatEng", "DR_TSNE_All"
             , "DR_GRP_Raw_Bin", "DR_GRP_Encoded", "DR_GRP_FeatEng", "DR_GRP_All"
             , "DR_SRP_Raw_Bin", "DR_SRP_Encoded", "DR_SRP_FeatEng", "DR_SRP_All"]

cols_core = dt_model[["ID", "y", "IsTrainTest"]].join(dt_model[cols_raw]).columns.values
dt_model_core = dt_model[cols_core]
dt_model_nonCore = dt_model.drop(cols_core, axis = 1)

list_featureSelect = []
for i in range(0, 100):
    # random sample
    rand_sample = sp_randint(round(len(list_cols) / 2), len(list_cols) + 1).rvs(1)[0]
    list_cols_random = random.sample(list_cols, rand_sample)
    print("list_cols_random", list_cols_random)
    regex_cols_random = "|".join(list_cols_random)
    dt_model_nonCore_sample = dt_model_nonCore.filter(regex = regex_cols_random)
    # join with random sample
    dt_model_sample = dt_model_core.join(dt_model_nonCore_sample)
    print("shape", dt_model_sample.shape)
    # X, y, ID
    X_train_all = dt_model_sample.loc[dt_model_sample["IsTrainTest"] == "train"].drop(["y", "IsTrainTest"], axis = 1)
    X_test = dt_model_sample.loc[dt_model_sample["IsTrainTest"] == "test"].drop(["y", "IsTrainTest"], axis = 1)
    y_train_all = dt_model_sample.loc[dt_model_sample["IsTrainTest"] == "train"].y.values
    y_test = dt_model_sample.loc[dt_model_sample["IsTrainTest"] == "test"].y.values
    ID_train_all = dt_model_sample.loc[dt_model_sample["IsTrainTest"] == "train"].ID.values
    ID_test = dt_model_sample.loc[dt_model_sample["IsTrainTest"] == "test"].ID.values
    # xgbDmatrix
    dmx_train_all = xgb.DMatrix(X_train_all, label = y_train_all)
    dmx_test = xgb.DMatrix(X_test)
    # base_score
    params_xgb["base_score"] = np.mean(y_train_all)
    # cv
    cv_xgb = xgb.cv(params_xgb, dmx_train_all
                    , num_boost_round = 2000
                    , nfold = 10
                    , feval = r_2, maximize = True, early_stopping_rounds = 50
                    , show_stdv = True, verbose_eval = 50)
    # record
    score = cv_xgb["test-score-mean"][cv_xgb.shape[0] - 1]
    sd = cv_xgb["test-score-std"][cv_xgb.shape[0] - 1]
    dict_featureSelect = {"score": score
                         , "sd": sd
                         , "nonCoreFeatures": list_cols_random}
    list_featureSelect.append(dict_featureSelect)

score_max = np.max([x["score"] for x in list_featureSelect])
cols_nonCore_max = [x["nonCoreFeatures"] for x in list_featureSelect if x["score"] == score_max][0]

regex_cols_nonCore_max = "|".join(cols_nonCore_max)

### Produce model

# dt_model_features = dt_model_core.join(dt_model_nonCore.filter(regex = regex_cols_nonCore_max))
# dt_model_features.shape

# remove ohe and targetMean
# dt_model_features = dt_model.drop(dt_model.filter(regex = "Encode_ohe|Encode_Freq|Encode_Binary|FeatEng_SumBin|single_train|dup_train").columns, axis = 1)
# dt_model_features.shape

# remove ohe
dt_model_features = dt_model.drop(dt_model.filter(regex = "Encode_ohe").columns, axis = 1)
dt_model_features.shape

# X, y, ID
X_train_all = dt_model_features.loc[dt_model_features["IsTrainTest"] == "train"].drop(["y", "IsTrainTest"], axis = 1)
X_test = dt_model_features.loc[dt_model_features["IsTrainTest"] == "test"].drop(["y", "IsTrainTest"], axis = 1)
y_train_all = dt_model_features.loc[dt_model_features["IsTrainTest"] == "train"].y.values
y_test = dt_model_features.loc[dt_model_features["IsTrainTest"] == "test"].y.values
ID_train_all = dt_model_features.loc[dt_model_features["IsTrainTest"] == "train"].ID.values
ID_test = dt_model_features.loc[dt_model_features["IsTrainTest"] == "test"].ID.values

X_train_all.shape

# xgbDmatrix
dmx_train_all = xgb.DMatrix(X_train_all, label = y_train_all)
dmx_test = xgb.DMatrix(X_test)

params_xgb["base_score"] = np.mean(y_train_all)

k = 10
# bin
bin_y = pd.qcut(y_train_all, k, labels = [i for i in range(1, k + 1)]).astype("int64")
# stratified kfold
skf = StratifiedKFold(n_splits = k, shuffle = True, random_state = 888)

score_skf_valid = 0
score_skf = 0
preds_skf_test = []
for i, (ind_train, ind_valid) in enumerate(skf.split(X_train_all, bin_y)):
    # X, y
    X_train, X_valid = X_train_all.iloc[ind_train], X_train_all.iloc[ind_valid]
    y_train, y_valid = y_train_all[ind_train], y_train_all[ind_valid]
    # xgb.DMatrix
    dmx_train_all = xgb.DMatrix(X_train, label = y_train)
    dmx_valid_all = xgb.DMatrix(X_valid, label = y_valid)
    ls_watch =  [(dmx_train_all, 'train'), (dmx_valid_all, 'eval')]
    
    model_xgb = xgb.train(params_xgb, dmx_train_all, evals = ls_watch
                          , num_boost_round = 5000
                          , feval = r_2, maximize = True, early_stopping_rounds = 50
                          , verbose_eval = 50
                         )
    
    preds_valid = model_xgb.predict(dmx_valid_all)
    score_skf_valid = r2_score(y_valid, preds_valid)
    print('Fold %d: Score %f'%(i, score_skf_valid))

    score_skf += score_skf_valid

score_skf /= k
print('=====================')

print( 'Final Score %f'%score_skf)

print('=====================')

# cv
cv_xgb = xgb.cv(params_xgb, dmx_train_all
                , num_boost_round = 5000
                , nfold = 10
                , feval = r_2, maximize = True, early_stopping_rounds = 50
                , show_stdv = True, verbose_eval = 50)

cv_xgb.tail()

# train
model_xgb = xgb.train(params_xgb, dmx_train_all, num_boost_round = cv_xgb.shape[0])

# importance
fig, ax = plt.subplots(figsize = (12,18))
xgb.plot_importance(model_xgb, max_num_features = 50, height = 0.8, ax = ax)
plt.show()

### Submit

# predict
y_test = model_xgb.predict(dmx_test)
y_test[:10]

# submit
dt_submit = pd.DataFrame({"ID": ID_test, "y": y_test})

dt_submit.to_csv("../../data/Mercedes_Benz_Greener_Manufacturing/submission/15_added_FI_2way_XOR.csv", index = False)
